pyproject.toml
[tool.poetry]
name = "llm-memory-vault"
version = "0.1.0"
description = "Local-first AI memory system aggregating LLM conversations into a searchable, secure knowledge base."
authors = ["Rich <rich@example.com>"]
readme = "README.md"
packages = [{include = "vault"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
typer = "^0.9.0"
pydantic = "^2.5.0"
cryptography = "^42.0.0"
argon2-cffi = "^23.1.0"
sqlalchemy = "^2.0.25"
qdrant-client = "^1.7.0"
ollama = "^0.1.6"
sentence-transformers = "^2.3.1"
presidio-analyzer = "^2.2.351"
pyyaml = "^6.0.1"
rich = "^13.7.0"
python-multipart = "^0.0.9"
uvicorn = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
vault = "vault.cli.main:app"
